22-12-26/Todo API1/requirement.txt
asgiref==3.6.0
Django==3.2.16
djangorestframework==3.13.1
pytz==2022.7
sqlparse==0.4.3
